+++ app.js
@@ -93,6 +93,8 @@ app.use(passport.initialize());
 app.use(passport.session());
 app.use(methodOverride('_method'));
 
+app.use()
+
 app.get('/create-profile',checkAuthenticated,(req,res) => {
     mysql.pool.query("CALL GetInstrumentsLevels()", [], (error, rows) => {
         if(error) {
@@ -125,7 +127,7 @@ app.get('/dashboard', checkAuthenticated, function (req, res, next) {
                 req.session.Profile = results[0]; //put profile in the session
 
                 context['profile'] = results;
-                getInstrumentsAndLevels(req, res, context, complete);
+                utils.getInstrumentsAndLevels(req, res, context, complete);
                 function complete() {
                     res.render('dashboard', context);
                 }
@@ -137,221 +139,26 @@ app.get('/dashboard', checkAuthenticated, function (req, res, next) {
     });
 });
 
-app.get('/dashboard/ads', checkAuthenticated, function (req, res, next) {
-    let callbackCount = 0;
-    let context = {
-        user: req.user
-    };
-    getInstrumentsAndLevels(req, res, context, complete);
-    getAds(req, res, context, complete);
-    function complete() {
-        callbackCount++;
-        if (callbackCount >= 3) {
-            res.send(context);
-        }
-    }
-});
-
-app.put('/dashboard/ads/enable', checkAuthenticated, function (req, res, next) {
-    mysql.pool.query("UPDATE Ads SET IsActive = ? WHERE AdKey = ?;", [req.body.IsActive, req.body.AdKey], (error, results) => {
-        if(error) {
-            res.write(JSON.stringify(error));
-            res.end();
-        } else {
-            res.send({ message: 'Successfully enabled ad.' });
-        }
-    });
-});
-
-app.delete('/dashboard/ads/delete', checkAuthenticated,(req, res, next) => {
-    mysql.pool.query("DELETE Ads, AdInstruments FROM Ads LEFT JOIN AdInstruments ON Ads.AdKey = AdInstruments.AdID WHERE Ads.AdKey = ?;",
-        [req.body.AdKey], (error, results) => {
-        if(error){
-            res.write(JSON.stringify(error));
-            res.end();
-        } else {
-            res.send({ message: 'Successfully deleted ad.' });
-        }
-    });
-});
-
-app.post('/dashboard/ads/edit', checkAuthenticated, (req, res, next) => {
-    try {
-        mysql.pool.getConnection(function (err, conn) {
-            if (err) throw (err);
-            //update the ad info
-            conn.query(
-                'UPDATE Ads SET Title = ?, Description = ?,  ZipCode = ?, LocationRadius = ?, DatePosted = NOW(), LastUpdated = NOW() WHERE AdKey = ? AND UserID = ?',
-                [req.body["ad-title"], req.body["ad-text"], req.body["ad-zip"], req.body["ad-radius"], req.body["ad-id"], req.user.UserKey],
-                function (err, rows) {
-                    if (err) {
-                        conn.release();
-                        res.write({ message: 'An error occurred when updating your ad: '.JSON.stringify(err) });
-                        res.end();
-                    }
-                    else if (rows.affectedRows === 1) //now that ad is updated, do any instruments updates
-                    {
-                        var adKey = req.body["ad-id"];
-
-                        //delete current instruments
-                        conn.query(`DELETE FROM AdInstruments WHERE AdId = ? `, [adKey],
-                            function (err, rows) {
-
-                                if (err) {
-                                    conn.release();
-                                    res.write({ message: 'An error occurred when updating your ad: '.JSON.stringify(err) });
-                                    res.end();
-                                } else {
-
-                                    //first format instrument/levelIDs sent in with form
-                                    var instruments = [];
-
-                                    var timestamp = new Date().toISOString().slice(0, 19).replace('T', ' '); //timestamp for create/lastupdated
-                                    for (i = 0; i <= 20; i++) { //set arbitary max of 20 instruments for now
-
-                                        if (Object.prototype.hasOwnProperty.call(req.body, "instruments[" + i + "][InstrumentKey]") &&
-                                            req.body["instruments[" + i + "][InstrumentKey]"] > 0) {
-
-                                            instruments.push([
-                                                req.body["instruments[" + i + "][InstrumentKey]"],
-                                                req.body["instruments[" + i + "][LevelKey]"],
-                                                req.body["instruments[" + i + "][Quantity]"],
-                                                adKey, timestamp, timestamp]);
-
-                                        }
-                                        else break;
-                                    }
-
-                                    ////add the instruments
-                                    conn.query(`INSERT INTO AdInstruments (InstrumentID, LevelID, Quantity, AdId, CreateDate, LastUpdated)  VALUES ?  `, [instruments],
-                                        function (err, rows) {
-                                            conn.release();
-
-                                            if (err) {
-                                                res.write({ message: 'An error occurred when updating your ad: '.JSON.stringify(err) });
-                                                res.end();
-                                            } else res.send({ success: 1, message: 'Successfully updated ad.' });
-
-                                        });
-                                }
-                            });
-                    }
-                    else {
-                        conn.release();
-                        res.send({ success: 0,  message: 'Ad not found.' });
-                    }
-                });
-        });
-    } catch (err) {
-        res.write(JSON.stringify(err));
-        res.end();
-    }
-});
-
-function getInstrumentsAndLevels(req, res, context, complete) {
-    mysql.pool.query("CALL GetInstrumentsLevels()", [], (error, rows) => {
-        if(error) {
-            throw(error);
-        }
-        context['instruments'] = rows[0];
-        context['levels'] = rows[1];
-        complete();
-    });
-}
-
-function getAds(req, res, context, complete) {
-    let sql = "SELECT `Value` FROM UserSettings WHERE UserID = ? AND SettingID = (SELECT SettingKey FROM Settings WHERE `Name` = 'AdSortOrder');"
-    mysql.pool.query(sql, [context.user.UserKey], (error, rows) => {
-        if (error) {
-            throw(error);
-        } else if (rows.length > 0) {
-            sql = "SELECT a.AdKey, a.Title, a.Description, a.ZipCode, a.LocationRadius, a.DatePosted, a.Deleted, " +
-                "a.DateCreated, a.LastUpdated, a.IsActive FROM Ads a WHERE a.UserID = ? ORDER BY " + rows[0]['Value'];
-            // set the sort value so that it can be read on refresh using the javascript in dashboard.js
-            context['sort'] = rows[0]['Value'];
-        } else {
-            sql = "SELECT a.AdKey, a.Title, a.Description, a.ZipCode, a.LocationRadius, a.DatePosted, a.Deleted, " +
-                "a.DateCreated, a.LastUpdated, a.IsActive FROM Ads a WHERE a.UserID = ? ORDER BY a.DatePosted DESC";
-            // set the sort value so that it can be read on refresh using the javascript in dashboard.js
-            context['sort'] = "a.DatePosted DESC";
-        }
-        mysql.pool.query(sql, [context.user.UserKey], (error, rows) => {
-            if (error) {
-                throw(error);
-            } else if (rows.length > 0) {
-                var ads = rows;
-                complete();
-                var ad_ids = jp.query(ads, "$..AdKey");
-                mysql.pool.query("SELECT ai.AdId, i.InstrumentKey, i.Instrument, l.LevelKey, l.Level, ai.Quantity\n" +
-                    "FROM AdInstruments ai\n" +
-                    "LEFT JOIN InstrumentLookup i ON i.InstrumentKey = ai.InstrumentID\n" +
-                    "LEFT JOIN LevelLookup l ON l.LevelKey = ai.LevelID\n" +
-                    "WHERE ai.AdID IN(?)\n" +
-                    "ORDER BY ai.AdID;", [ad_ids], (error, rows) => {
-                    if (error) {
-                        throw(error);
-                    } else if (rows.length > 0) {
-                        for (let ad of ads) {
-                            ad['instruments'] = rows.filter(row => row.AdId == ad['AdKey']);
-                        }
-                        context['current_ads'] = ads.filter(ad => ad.IsActive === 1);
-                        context['has_current_ads'] = (context['current_ads'].length > 0);
-                        context['prev_ads'] = ads.filter(ad => ad.IsActive === 0);
-                        context['has_prev_ads'] = (context['prev_ads'].length > 0);
-                        complete();
-                    } else {
-                        complete();
-                    }
-                });
-            } else {
-                complete();
-                complete(); //need to add another call to callback so we can render dashboard
-                context['has_current_ads'] = false;
-                context['has_prev_ads'] = false;
-            }
-        });
-    });
-}
-
 // FIXME: This route will need to get all the users that match the ad criteria.
-app.get('/search-results', checkAuthenticated, function(req, res, next) {
+app.get('/search-results', utils.checkAuthenticated, function(req, res, next) {
     // Do some sort of awful select with joins, then render (nav bar needs the user's profile to render):
     res.render('search-results', { profile: true });
 });
 
-app.post('/adSortOrder', checkAuthenticated, function (req, res, next) {
-    // checks whether the UserID exists in the table
-    mysql.pool.query("SELECT UserID FROM UserSettings WHERE UserID = ?;", [req.user.UserKey], (error, results) => {
-           if (results.length == 0) {
-                //there is no sortOrder in table for this user so we create a new one 
-                mysql.pool.query("INSERT INTO UserSettings (UserID, SettingID, `Value`) VALUES (?, (SELECT SettingKey FROM Settings WHERE `Name` = 'AdSortOrder'), ?);", [req.user.UserKey, req.body.sortOrder], (error, results) => {
-                   return res.redirect('\dashboard');
-                });
-            } else {
-                // a sortOrder that is not the default exists already so we just update the Value in the backend
-                mysql.pool.query("UPDATE UserSettings SET `Value` = ? WHERE UserID = ?;", [req.body.sortOrder, req.user.UserKey], (error, results) => {
-                   return res.redirect('\dashboard');
-                });
-            } 
-    });
-
-});
-
-
 //any page requiring NOT authentication needs to run checkNotAuthenticated first
 //landing page does not need authentication, in fact we do not allow logged in users to access
-app.get('/', checkNotAuthenticated, function (req, res, next) {
+app.get('/', utils.checkNotAuthenticated, function (req, res, next) {
     res.render('landing', { successRegistration: req.flash('successRegistration'), errorRegistration: req.flash('errorRegistration')});
 });
 
 //passport.js will handle login page
-app.post('/login', checkNotAuthenticated, passport.authenticate('local', {
+app.post('/login', utils.checkNotAuthenticated, passport.authenticate('local', {
     successRedirect: '/dashboard',
     failureRedirect: '/',
     failureFlash: true
 }));
 
-app.post('/register', checkNotAuthenticated, async function (req, res) {
+app.post('/register', utils.checkNotAuthenticated, async function (req, res) {
     try {
         const hashedPassword = await bcrypt.hash(req.body.password, 10); //encrypt the password entered by the user
 
@@ -390,7 +197,7 @@ app.delete('/logout', (req, res) => {
     res.redirect('/'); //on logout, redirect to landing page
 });
 
-app.get('/profile',checkAuthenticated,(req,res,next) => {
+app.get('/profile',utils.checkAuthenticated,(req,res,next) => {
     try {
         mysql.pool.query('CALL GetProfile(?)', [req.user.UserKey], function(err, rows) {
             if(err) {
@@ -406,7 +213,7 @@ app.get('/profile',checkAuthenticated,(req,res,next) => {
                     }
                 };
 
-                getInstrumentsAndLevels(req, res, context, complete);
+                utils.getInstrumentsAndLevels(req, res, context, complete);
 
                 function complete() {
                     res.render('profile', context);
@@ -421,7 +228,7 @@ app.get('/profile',checkAuthenticated,(req,res,next) => {
 });
 
 // saves the info that is located in the Profiles header and returns true if update was successful
-app.put('/profile/header', checkAuthenticated,(req, res, next) => {
+app.put('/profile/header', utils.checkAuthenticated,(req, res, next) => {
     try {
         mysql.pool.query(
             'UPDATE Profiles SET ZipCode = ?, ArtistName = ?, LastUpdated = NOW(), LookingForWork = ? WHERE UserID = ?',
@@ -441,7 +248,7 @@ app.put('/profile/header', checkAuthenticated,(req, res, next) => {
 });
 
 // saves the info that is located in the Profiles about/bio and returns true if update was successful
-app.put('/profile/about', checkAuthenticated,(req, res, next) => {
+app.put('/profile/about', utils.checkAuthenticated,(req, res, next) => {
     try {
         mysql.pool.query(
             'UPDATE Profiles SET Bio = ?, LastUpdated = NOW() WHERE UserID = ?',
@@ -461,7 +268,7 @@ app.put('/profile/about', checkAuthenticated,(req, res, next) => {
 });
 
 // saves the info that is located in the Profiles website/social section and returns true if update was successful
-app.put('/profile/website', checkAuthenticated,(req, res, next) => {
+app.put('/profile/website', utils.checkAuthenticated,(req, res, next) => {
     try {
         mysql.pool.query(
             'UPDATE Profiles SET Website = ?, LastUpdated = NOW() WHERE UserID = ?',
@@ -481,7 +288,7 @@ app.put('/profile/website', checkAuthenticated,(req, res, next) => {
 });
 
 //updates the profile instruments. Called from within the profile view page.
-app.post('/profile/instruments', checkAuthenticated, (req, res, next) => {
+app.post('/profile/instruments', utils.checkAuthenticated, (req, res, next) => {
     try {
         mysql.pool.getConnection(function (err, conn) {
             if (err) throw (err);
@@ -537,7 +344,7 @@ app.post('/profile/instruments', checkAuthenticated, (req, res, next) => {
     }
 });
 
-app.delete('/profile/worksamples/music',checkAuthenticated,(req, res, next) => {
+app.delete('/profile/worksamples/music',utils.checkAuthenticated,(req, res, next) => {
     try {
         mysql.pool.query('DELETE FROM WorkSamples WHERE ProfileID=? AND SampleKey=? AND SampleType="Music"',
             [req.session.ProfileID, req.body.sampleKey],
@@ -555,7 +362,7 @@ app.delete('/profile/worksamples/music',checkAuthenticated,(req, res, next) => {
     }
 });
 
-app.put('/profile/worksamples/music',checkAuthenticated,(req, res, next) => {
+app.put('/profile/worksamples/music',utils.checkAuthenticated,(req, res, next) => {
    try {
        mysql.pool.query('UPDATE WorkSamples SET SampleLocation=? WHERE SampleKey=? AND ProfileID=? AND SampleType="Music"',
            [req.body.workSampleTextInput, req.body.id, req.session.ProfileID],
@@ -573,7 +380,7 @@ app.put('/profile/worksamples/music',checkAuthenticated,(req, res, next) => {
    }
 });
 
-app.post('/profile/worksamples/music',checkAuthenticated,(req, res, next) => {
+app.post('/profile/worksamples/music',utils.checkAuthenticated,(req, res, next) => {
     try {
         mysql.pool.query('INSERT INTO WorkSamples (ProfileID, SampleLocation, SampleType) VALUES (?, ?, "Music")',
             [req.session.ProfileID, req.body.workSampleTextInput],
@@ -591,7 +398,7 @@ app.post('/profile/worksamples/music',checkAuthenticated,(req, res, next) => {
     }
 });
 
-app.delete('/profile/worksamples/video',checkAuthenticated,(req, res, next) => {
+app.delete('/profile/worksamples/video',utils.checkAuthenticated,(req, res, next) => {
     try {
         mysql.pool.query('DELETE FROM WorkSamples WHERE ProfileID=? AND SampleKey=? AND SampleType="Video"',
             [req.session.ProfileID, req.body.sampleKey],
@@ -609,7 +416,7 @@ app.delete('/profile/worksamples/video',checkAuthenticated,(req, res, next) => {
     }
 });
 
-app.put('/profile/worksamples/video',checkAuthenticated,(req, res, next) => {
+app.put('/profile/worksamples/video',utils.checkAuthenticated,(req, res, next) => {
     try {
         mysql.pool.query('UPDATE WorkSamples SET SampleLocation=? WHERE SampleKey=? AND SampleType="Video"',
             [req.body.workSampleTextInput, req.body.id],
@@ -627,7 +434,7 @@ app.put('/profile/worksamples/video',checkAuthenticated,(req, res, next) => {
     }
 });
 
-app.post('/profile/worksamples/video',checkAuthenticated,(req, res, next) => {
+app.post('/profile/worksamples/video',utils.checkAuthenticated,(req, res, next) => {
     try {
         mysql.pool.query('INSERT INTO WorkSamples (ProfileID, SampleLocation, SampleType) VALUES (?, ?, "Video")',
             [req.session.ProfileID, req.body.workSampleTextInput],
@@ -646,7 +453,7 @@ app.post('/profile/worksamples/video',checkAuthenticated,(req, res, next) => {
 });
 
 // saves the info that is located in the Profiles table and returns true if the update was successful
-app.post('/profile/basic',checkAuthenticated,(req, res, next) => {
+app.post('/profile/basic',utils.checkAuthenticated,(req, res, next) => {
     try {
         mysql.pool.query(
             'UPDATE Profiles SET ZipCode = ?, Phone = ?, Website = ?, LookingForWork = ?, ArtistName = ?, LastUpdated = NOW() WHERE UserID = ?',
@@ -665,7 +472,7 @@ app.post('/profile/basic',checkAuthenticated,(req, res, next) => {
     }
 });
 
-app.post('/profile/basic/create', checkAuthenticated, (req, res, next) => {
+app.post('/profile/basic/create', utils.checkAuthenticated, (req, res, next) => {
     try {
         mysql.pool.getConnection(function (err, conn) {
             if (err) throw (err);
@@ -735,7 +542,7 @@ app.post('/profile/basic/create', checkAuthenticated, (req, res, next) => {
     }
 });
 
-app.get('/profile/instruments',checkAuthenticated,(req, res, next) => {
+app.get('/profile/instruments',utils.checkAuthenticated,(req, res, next) => {
     try {
         mysql.pool.query(
             'SELECT InstrumentKey, Instrument, SearchTerm FROM InstrumentLookup',
@@ -754,7 +561,7 @@ app.get('/profile/instruments',checkAuthenticated,(req, res, next) => {
     }
 });
 
-app.get('/profile/levels',checkAuthenticated,(req, res, next) => {
+app.get('/profile/levels',utils.checkAuthenticated,(req, res, next) => {
     try {
         mysql.pool.query(
             'SELECT LevelKey, Level FROM LevelLookup',
@@ -774,7 +581,7 @@ app.get('/profile/levels',checkAuthenticated,(req, res, next) => {
 });
 
 // inserts an instrument and associated level and returns true if the insert was successful
-app.post('/profile/instrument/add', checkAuthenticated, (req, res, next) => {
+app.post('/profile/instrument/add', utils.checkAuthenticated, (req, res, next) => {
     try {
         mysql.pool.query(
             'INSERT INTO ProfileInstruments (ProfileID, InstrumentID, LevelID, CreateDate) VALUES (?, ?, ?, NOW())',
@@ -793,7 +600,7 @@ app.post('/profile/instrument/add', checkAuthenticated, (req, res, next) => {
     }
 });
 
-app.get('/profile/worksamples', checkAuthenticated, (req, res, next) => {
+app.get('/profile/worksamples', utils.checkAuthenticated, (req, res, next) => {
     try {
         mysql.pool.query(
             'SELECT ProfileKey FROM Profiles WHERE UserID = ?',
@@ -821,7 +628,7 @@ app.get('/profile/worksamples', checkAuthenticated, (req, res, next) => {
 });
 
 // updates an instrument and associated level and returns true if the update was successful
-app.post('/profile/instrument/update',checkAuthenticated,(req, res, next) => {
+app.post('/profile/instrument/update',utils.checkAuthenticated,(req, res, next) => {
     try {
         mysql.pool.query(
             'UPDATE ProfileInstruments SET LevelID = ?, LastUpdated = NOW() WHERE ProfileID = ? AND InstrumentID = ?',
@@ -842,7 +649,7 @@ app.post('/profile/instrument/update',checkAuthenticated,(req, res, next) => {
 
 //for storing data from ad creation, use req.body[] because otherwise it is read in as a subtraction
 //added form action and method, also changed from datalist to regular select.
- app.post('/dashboard/ads/create', checkAuthenticated, (req, res, next) => {
+ app.post('/dashboard/ads/create', utils.checkAuthenticated, (req, res, next) => {
      try {
          mysql.pool.getConnection(function (err, conn) {
              if (err) throw (err);
@@ -897,12 +704,12 @@ app.post('/profile/instrument/update',checkAuthenticated,(req, res, next) => {
      }
 });
 
-app.get('/matches',checkAuthenticated,(req, res, next) => {
+app.get('/matches',utils.checkAuthenticated,(req, res, next) => {
     // Get all matches in the matches table, then:
     res.render('matches', { profile: true });
 });
 
-app.get('/matches/pending',checkAuthenticated,(req, res, next) => {
+app.get('/matches/pending',utils.checkAuthenticated,(req, res, next) => {
    try {
        let sql = 'SELECT * FROM Matches WHERE Accepted = 0 AND MatchedProfileID = ?';
        mysql.pool.query(sql, [req.session.ProfileID], function (err, result) {
@@ -916,24 +723,6 @@ app.get('/matches/pending',checkAuthenticated,(req, res, next) => {
        res.redirect(utils.errorRedirect('/matches/pending', 'An unexpected error occurred retrieving your matches'));
    }
 });
-
-//for pages accessible by authenticated users only
-function checkAuthenticated(req, res, next) {
-    if (req.isAuthenticated()) {
-        //TODO add check to see if user has created profile. If not direct them to create a profile
-        return next();
-    }
-
-    res.redirect('/'); //if not authenticated, bump to landing page
-}
-
-//for pages accessible by non-authenticated users only
-function checkNotAuthenticated(req, res, next) {
-    if (req.isAuthenticated()) {
-        return res.redirect('/dashboard'); //if authenticated, bump to dashboard page
-    }
-    next();
-}
 
 // start app
 app.listen(port, function(){
